Add LineAcceptanceTest to divide from StationAcceptanceTest Simplify StationAcceptanceTest
